refactor filenames to comply with code guidelines
